Move `interfaces/jsonapi/*` folder into `infrastructure/serializers/jsonapi/response-objects`
