+++ api/lib/application/assessments/assessment-controller.js
@@ -2,7 +2,7 @@ const { AssessmentEndedError } = require('../../domain/errors');
 const usecases = require('../../domain/usecases');
 const events = require('../../domain/events');
 const logger = require('../../infrastructure/logger');
-const JSONAPI = require('../../interfaces/jsonapi');
+const JSONAPI = require('../../infrastructure/serializers/jsonapi/response-objects');
 const assessmentRepository = require('../../infrastructure/repositories/assessment-repository');
 const assessmentSerializer = require('../../infrastructure/serializers/jsonapi/assessment-serializer');
 const challengeSerializer = require('../../infrastructure/serializers/jsonapi/challenge-serializer');
