+++ api/lib/infrastructure/serializers/jsonapi/response-objects/bad-request-error.js
