+++ api/lib/infrastructure/serializers/jsonapi/response-objects/forbidden-error.js
