+++ api/lib/infrastructure/serializers/jsonapi/response-objects/internal-error.js
