+++ api/lib/infrastructure/serializers/jsonapi/response-objects/not-found-error.js
