+++ api/lib/infrastructure/serializers/jsonapi/response-objects/unauthorized-error.js
