+++ mon-pix/app/components/feedback-panel.js
@@ -1,7 +1,7 @@
 import { classNames } from '@ember-decorators/component';
 import { action, computed } from '@ember/object';
 import { inject as service } from '@ember/service';
-import { isEmpty } from '@ember/utils';
+import { isEmpty } from '@ember/response-objects';
 import $ from 'jquery';
 import config from 'mon-pix/config/environment';
 import Component from '@ember/component';
